[feature][better_cleanup] Extracting the shelf stuff to its own service
