+++ src/main/kotlin/br/org/texugo/butterbot/rockstock/service/RockStockService.kt
@@ -1,100 +1,33 @@
 package br.org.texugo.butterbot.rockstock.service
 
-import br.org.texugo.butterbot.rockstock.CompressionException
 import br.org.texugo.butterbot.rockstock.DocumentNotFoundException
 import br.org.texugo.butterbot.rockstock.MoveFileException
 import br.org.texugo.butterbot.rockstock.OpsyException
 import br.org.texugo.butterbot.rockstock.data.Document
+import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.io.File
-import java.io.FileInputStream
-import java.io.FileOutputStream
 import java.nio.file.Files
 import java.nio.file.Paths
-import java.security.MessageDigest
 import java.util.concurrent.CompletableFuture
-import java.util.zip.ZipEntry
-import java.util.zip.ZipOutputStream
-import kotlin.experimental.and
 
 class RockStockService (
 
         val bufferSize : Int,
         private val _directory : String,
-        val tempDirectory : String
+        private val tempDirectory : String
 
 ) {
 
-    val log = LoggerFactory.getLogger(RockStockService::class.java)
-
     companion object {
+
+        val LOG : Logger = LoggerFactory.getLogger(RockStockService::class.java)
+
         const val DIGEST_TYPE = "SHA-1"
     }
 
-    val directory = _directory
+    private val directory = _directory
         get() = if (field.endsWith(File.separator)) field else "$field${File.separator}"
-
-    @Throws(OpsyException::class)
-    private fun shelfilize (tempDocument : Document) : Document {
-
-        return try {
-
-            log.debug("Shelfilizing temporary document ${tempDocument.canonicalPath}")
-
-            val shelfilizedDocument = Document(
-                    "$directory${FileService.extractDocumentFilename(tempDocument)}"
-            )
-
-            val shelfilizedDirectoryPath = Paths.get(FileService.extractDocumentDirectory(shelfilizedDocument))
-
-            when {
-                Files.notExists(shelfilizedDirectoryPath) -> Files.createDirectories(shelfilizedDirectoryPath)
-            }
-
-            val tempDocumentPath = Paths.get(tempDocument.canonicalPath)
-            val shelfilizedDocumentPath = Paths.get(shelfilizedDocument.canonicalPath)
-
-            when {
-                Files.exists(shelfilizedDocumentPath) -> Files.delete(shelfilizedDocumentPath)
-            }
-
-            Files.move(tempDocumentPath, shelfilizedDocumentPath).let {
-
-                when {
-                    Files.exists(tempDocumentPath) -> Files.delete(tempDocumentPath)
-                }
-
-                // Returning the shelfilized document if the move
-                // operation went well
-                it?.let {
-
-                    log.debug("Document ${shelfilizedDocument.canonicalPath} successfully shelfilized")
-
-                    return shelfilizedDocument
-
-                }
-
-            }
-
-            log.warn("Huge probability that the document ${
-                tempDocument.canonicalPath
-            } wasn't moved to ${
-                shelfilizedDocument.canonicalPath
-            }")
-
-            throw MoveFileException(tempDocument.canonicalPath, shelfilizedDocument.canonicalPath)
-
-        } catch (err : Exception) {
-
-            val message = "Opsy! An error while \"shelfilizing\" the document ${tempDocument.canonicalPath}"
-
-            log.error(message, err)
-
-            throw OpsyException(message, err)
-
-        }
-
-    }
 
     @Throws(OpsyException::class)
     fun stockIt (originalDocument : Document) : Document {
@@ -103,17 +36,23 @@ class RockStockService (
 
             return CompletableFuture.supplyAsync {
 
-                log.debug("Stocking the document ${originalDocument.canonicalPath}")
+                LOG.debug("Stocking the document ${originalDocument.canonicalPath}")
 
                 CompressionService.compress(originalDocument, this.tempDirectory)
 
-            }.thenApplyAsync(this::shelfilize).join()
+            }.thenApplyAsync { compressedDocument ->
+
+                LOG.debug("Shelfilizing the document ${compressedDocument.canonicalPath} in ${this.directory}")
+
+                ShelfService.shelfilize(compressedDocument, this.directory)
+
+            }.join()
 
         } catch (err : Exception) {
 
             val message = "Opsy! An error while stocking the document ${originalDocument.canonicalPath}"
 
-            log.error(message, err)
+            LOG.error(message, err)
 
             throw OpsyException(message, err)
 
@@ -128,7 +67,7 @@ class RockStockService (
 
             return CompletableFuture.supplyAsync {
 
-                log.debug("Rocking the document $id")
+                LOG.debug("Rocking the document $id")
 
                 val filePath = "$directory$id${
                     CompressionService.EXTENSION_SEPARATOR
@@ -136,7 +75,7 @@ class RockStockService (
                     CompressionService.EXTENSION_COMPRESSION
                 }"
 
-                log.debug("Calculated path $filePath for document $id")
+                LOG.debug("Calculated path $filePath for document $id")
 
                 val file = File(filePath)
 
@@ -154,7 +93,7 @@ class RockStockService (
 
             val message = "Opsy! Some error retrieving the document $id"
 
-            log.error(message, err)
+            LOG.error(message, err)
 
             throw OpsyException(message, err)
